chore: move gaugeStyles to appropriate folder
