+++ giraffe/src/style/gaugeStyles.ts
